Fix Character count's JSDOM tests
Now we throw if the textarea description is missing, that element needs to be in the DOM as well.
Our use of `document.getElementById` also meant that the elements need to be added to the `document`.
This meant that we also need some cleanup between tests as Jest JSDOM doesn't handle that between tests of a same file.
Finally, we also needed to use `textContent` rather than `innerText` as[JSDOM does not support the later](https://github.com/jsdom/jsdom/issues/1245),
and we were only [using it for IE8](https://github.com/alphagov/govuk-frontend/pull/2980), which we no longer need to support.