Mark PNG icons as maskable
The `maskable` keyword tells Android devices that the icon has been designed with a circular, central safe area, and that Android launchers can
safely crop the edges of the icon instead of shrinking the icon down:
https://css-tricks.com/maskable-icons-android-adaptive-icons-for-your-pwa/

Despite the name, these are not 'masks' in the same way as the Safari icon. The Safari icon is an alpha mask, which is indicated using the
`monochrome` keyword instead.