Add organisation brand colour for Department for Business Innovation and Skills
Department for Business Innovation and Skills has been broken into
several departments. We've been asked to update the branding class on
Whitehall from department-for-business-innovation-skills to
department-for-energy-security-and-net-zero.

In order to make sure there's no issues with backwards issues we're adding
a new org brand colour, rather than renaming the existing one.