Remove X-UA-Compatible meta tag
This meta tag ensured that older versions of Internet Explorer (IE8 onwards) used the most recent renderer and rendering rules available to
them.

IE11 does not require this meta tag. It deprecates the X-UA-Compatible meta tag and now uses the most
recent renderer automatically if the doctype is `<!DOCTYPE html>`.

Both the EdgeHTML and Chromium versions of Edge ignore the meta tag entirely.

https://learn.microsoft.com/en-us/previous-versions/windows/internet-explorer/ie-developer/dev-guides/bg182625(v=vs.85)#document-mode-changes