Remove ‘details polyfill’ example from review app
This example was used to test the JavaScript polyfill for the details component on a `<div>` to ensure that it added the same functionality as the native `<details>` element.

We removed that JavaScript polyfill in #3766.

This example no longer functions correctly (it renders something that looks like the details component, but clicking the summary does not open or close) and serves no purpose, so we can remove it.