[SPIKE] Handle component JS not loading correctly
At the minute, components can end up in a broken state if the user has JavaScript enabled but for whatever reason the script for a component fails to execute.

This is because we use the `js-enabled` class on the body to style a component based on the assumption that the JavaScript will execute, including for example hiding content in accordions and tabs. If the JS fails to run, this can result in components ‘failing closed’ or appearing to be interactive without doing anything.

Inspired by a comment from @colinrotherham last year I arrived at a potential solution:

> Could even have a “smoke and mirrors” class that we temporarily add with a setTimeout? This is the maximum time the Accordion should stay closed to avoid layout shift

The way it works is:

- In the inline script immediately after the opening `<body>` tag, we also add a new `govuk-frontend-loading` class to the body element. We also set up an event listener which removes the loading class when the `DOMContentLoaded` event fires (after deferred scripts have downloaded and executed).
- When the component is instantiated, we add a `--ready` modifier class to the `$module` (unless one of the various guards in the constructor throws)
- The CSS for the component is updated to e.g. only hide the accordion content whilst the loading class is present on the body _or_ the ready class is present on the module.

This means that one of two things should happen:

1. The JavaScript executes successfully, adding the `--ready` modifier. When the `loading` class is removed from the body, the element is styled
2. For whatever reason, the JavaScript fails to run. When the `loading` class is removed, the component falls back to the no-JS experience

Non-exhaustive list of questions still to answer:

- Are there any situations where the `DOMContentLoaded` event listener wouldn’t end up being called?
- Does this cause an overhead in terms of style invalidation / re-calcuation when we remove the class from the body?
- How does this work for components that are added to the page dynamically, for example as part of an SPA?