Remove `aria-labelledby` from accordion sections
Adding `aria-labelledby` to a `<div>` with no explicit role is explicitly disallowed by the 'ARIA in HTML' specification [1] and in our testing provides little or no benefit to users, including users of assistive technologies [2].

We’ve explored using `<section>` elements for each accordion section (for which a label would make sense), however best practise is to limit the use of region landmarks ‘as too many landmarks can dilute their usefulness in quickly navigating to important areas of a page’ [3].

We don’t have clear evidence that marking accordion sections up as regions is useful, so more research is needed in this area.

There has also been a suggestion that we could use `<article>` elements, but it would depend on how the accordion is being used. [4]

We may still want to pursue these options in the future, but until we make a decision on whether the sections should have a semantic role remove the `aria-labelledby` attribute – it’s trivial to reintroduce it later.

[1]: https://www.w3.org/TR/html-aria/#docconformance-naming
[2]: https://github.com/alphagov/govuk-frontend/issues/2472#issuecomment-1766282547
[3]: https://www.scottohara.me/blog/2018/03/03/landmarks.html#region
[4]: https://github.com/alphagov/govuk-frontend/issues/2472#issuecomment-1878807875