Omit nav when header navigation is an empty array
Ensure that the `navigation` param of the header is non-empty rather than truthy. This avoids a scenario where the wrapper, nav and menu button are outputted with an empty navigation `<ul>` when the `navigation` param is an empty array (`[]`):

```html
<div class="govuk-header__content">
  <nav aria-label="Menu" class="govuk-header__navigation">
    <button type="button" class="govuk-header__menu-button govuk-js-header-toggle" aria-controls="navigation" hidden>Menu</button>
    <ul id="navigation" class="govuk-header__navigation-list"></ul>
  </nav>
</div>
```

This issue is hard to spot as there is no visual change on tablet breakpoints and above – only on the mobile breakpoint where the menu button is visible (but clicking it only changes the arrow direction).