Only use pre-built fixtures when running in CI
Using `options?.fixture?.html` provides a neat speed optimisation when running the tests on CI (where we already rendered all of the fixtures to HTML and they will not change).

However, for local development, we currently only update the fixtures based on watching the YAML files. This means they are not updated when the Nunjucks template changes, so tests can run on outdated versions of templates, which is confusing and potentially misleading.

We discussed additionally watching for changes to the Nunjucks templates as well, however there is still a race condition when running tests (especially as the fixture generation currently takes ~1.6-1.8s).

For now, only use this optimisation on CI where we can guarantee we’re testing against an up to date build.

Co-authored-by: Colin Rotherham <colin.rotherham@digital.cabinet-office.gov.uk>