Reset component state on pageshow event
If a user filled out a password input and then navigated away from the page via a link, it is possible to use the browser back button to return to the page with the password field populated. It could then be viewed and copied using the show button. This could potentially be used by bad actors to obtain user passwords.

This code now blanks the input and resets the component state if the page is returned to, unless an explicit value has been set by the developer.

Athough we cannot guarantee complete safety in this scenario (as the attacker evidently has obtained some level of physical or remote access already), it may make this attack vector slightly more difficult to accomplish.