Iterate on dataset keys ordered as strings
This ensures that deeper keys (like `i18n.messageKey.other`) get processed after shallower ones (like `i18n.messageKey`) and override any existing value with a new object.

This sounds a legitimate reason for throwing, but would it be a breaking change to do so? Without throwing the behaviour is the same as with the flattened version we currently use