Remove disableFormSubmitCheck option
Discussions with the contributor previously revealed that this was added defensively 'just in case'. As far as we are aware it hasn't actually been used anywhere (both via anecdotal evidence and the only GitHub search results being spec tests), indicating it has little to no actual user need or demand.

In the interests of keeping things simple, we're going to remove this from our initial version.