Delete redundant test and example
This test is a little clunky and doesn’t have any assertions.

The same logic is covered by newer tests in src/govuk/all.jsdom.test.js (‘only initialises components within a given scope’) so we don’t need them for coverage.

The example has little value outside of the scope of the test, so remove that too.