Make status announcement assertive
Changes the aria-live attribute of the hidden status element from "polite" to "assertive".

This is intended to rectify an issue with VoiceOver on macOS, where the announcement would be spoken over by the changing text and aria-label of the toggle button. In this case, notifying the user that their password is visible (or not) feels like an important piece of information that should take priority.