Fix password input button scaling
Fixes an issue where the password input toggle button would stretch to fill any space not used by the input, such as if the input has a restricted max width.

I think the `flex-grow` property may have been a leftover from earlier in the component's development, as it doesn't appear to be needed now.

Adds a review app example to help keep this in check in future.