Fix failing accessibility test
Remove the 'with no options set' example from being tested as part of the standard accessibility tests.

This is because the accessibility test tries to initialise the component's JS, but there is no HTML being rendered to initialise it against, returning an ElementError that Jest considers to be a test failure.