Replace custom Sass logger w/ silenceDeprecations
Use the new `silenceDeprecations` feature introduced in Sass 1.74.0 and remove the custom logger that was implemented to filter them out.

The logger was introduced in 81f59d9 and ec547be as part of #3164.

As mentioned in the first commit message, the output otherwise exactly matches the log output from Dart Sass. I’ve also manually verified this by comparing the output from the custom logger with an empty array for `supressed` and using the default logged with an empty array for `silenceDeprecations`.

There’s nothing else using `chalk` which was introduced as a dependency in 81f59d9 so uninstall that too.