Make Dependabot ignore `eslint` & `eslint-*`
Dependabot keeps timing out before completing and on inspection of the
logs, the latest package it attempts to install is `eslint-plugin-n`.
This commit makes Dependabot ignore the eslint packages until we are
able to update them all, as this is currently blocked by not all the
plugins supporting the latest version of the `eslint` config.