Remove redundant tests
There’s already an expectation for the default text of ‘Back’ in the ‘renders the default example with an anchor, href and text correctly’ test.

The template has no option for inverse colours – the example just passes an additional class, so the test has no value over the existing test for additional classes.