Rename breadcrumbs template tests to .jsdom.test.js
The next commit will refactor the tests to make use of jsdom which requires us to update the filename so that we get the jsdom test environment.

Doing this in two commits so that git shows the changes to the file rather than treating it as an addition and a deletion.