Analyse versions between diff range
Results of me and Romaric pairing.

This change checks the repo's lockfile, gets the latest and oldest version from the commit diff and uses semver to check if the range includes a crown update.

Next steps are to analyse the commit range once we've established that the range contains a crown change and establish a way to step through commit history when we can't find a crown change.

We noted there's a risk of the github API rate limit becoming exceeded very quickly if we run this against the full list of 1280 repos so we need to think about ways to mitigate this with throttling or even manually stopping.