Define additional screenshots in component YAML
Rather than hardcoding the list of additional examples to screenshot in the task, add a new `screenshot` boolean to the component YAML (and generated fixtures.yaml) which determines whether the example should be included in the screenshots or not.

Preserve the existing list of screenshotted examples by setting `screenshot: true` in the YAML for those examples.

Remove special treatment of examples named ‘inverse’ but retain their screenshots by also setting `screenshot: true`.