Remove root `src` directory
This directory was left to facilitate the transition of contributors
after [we re-organised our repository in May last year].

The symlinks inside it cause a little trouble with VSCode,
offering two paths to reach a file, each sometime behaving a little
differently in terms of linting/type-checking on save.

With our new structure in place for more than a year now,
it's time to say goodbye to that directory and only access sources through `packages`.