Avoid unnecessary replacement
There is no point replacing the existing `$summary` with the `$summarySpan` [^1] as we end up moving the `$summarySpan` into the `$button` on the next line anyway.

All we really need to do is remove the existing summary, which is much simpler. Because we don’t need to traverse to the parent we can also simplify the guard for this block of code by removing the check for `$summary.parentNode`.

[^1]: Note that the argument order for `Node.replaceChild` may not be what you expect (`replaceChild(newChild, oldChild)`)