Bump minimum Safari version in Grade C to Safari 11
The summary of browser grades include Safari 10.1 (macOS) and 10.3 (iOS) as examples of Grade C browsers.

However in the Grade C section of the document it's mentioned that Safari 10.1 will not run our JavaScript as (although it technically supports `type="module"`) it fails the prerequisite support tests present in Frontend. That section lists Safari 11 (macOS) as being the minimum for Grade C instead.

This is confusing, as different parts of the document use different Safari versions as the minimum supported version for Grade C. It also does not mention that Safari 10.3 (iOS) is subject to the same issues as Safari 10.1 (macOS).

This commit updates the summary at the top of the document and in the Grade C section to use Safari 11 as the minimum supported version, as this is what appears to be the case to a person quickly glancing at the document in reference. The Grade C section has been updated to make it clear that, while Safari 10.1 (macOS) and Safari 10.3 (iOS) are technically included, they do not actually initialise any of Frontend's components due to failing other prerequisites outside of having support for modules.