Fix navigation extending off-screen in IE11
IE11 is tripping over the use of flex-wrap and is instead presenting all links as a single, horizontally scrolling row.

This commit reverts the use of flexbox on IE11 (and IE10) specifically, falling back to the previous inline-block behaviour. This means that the previous accessibility issues are now present in IE11 again, but having the navigation wrap seems like a higher priority usability issue to address.