Move `govuk-frontend` ESLint config to `@govuk-frontend/eslint-config`
Moves the `.eslintrc.js` file usit `git mv` to keep the history of the file.
Adding back a `.eslintrc.js` file in the same commit to keep ESLint working
would confuse Git, though. Instead, we can take advantage of the `.cjs` extension,
which will be understood just as well by ESLint, but is a different file from Git's perspective.