Add deprecated JavaScript files to package build
Because they're not imported by `all.mjs` deprecated JavaScript files were not included in the package (in which they need to remain until the next breaking release).

This adds a module to store the list of deprecated JavaScript files and helper functions that are then used:
- in the Gulp configuration, to add a new task building each deprecated file individually
- in the Rollup configuration, to prevent the bundled output of the deprecated files as all we want is for them to be in the package as a module

Add new section to documentation `Managing Change` which details the
process of marking a JS file for deprecation.

Co-authored-by: Patrick Cartlidge <patrick.cartlidge@digital.cabinet-office.gov.uk>
Co-authored-by: Brett Kyle <brett.kyle@digital.cabinet-office.gov.uk>
Co-authored-by: seaemsi <50173207+seaemsi@users.noreply.github.com>