Fixes issue in Safari where dropzone was disappearing
It happened when dragging from the button to the span or the opposite,
because Safari does not fill the `relatedTarget` on `dragleave`, making
our code believe user had left the window.

To accomodate for that, we use `dragenter` to also hide the drop zone
when entering an element that's not the wrapper.

This may still leave a gap where the component is at the edge of the viewport,
either because of scrolling or in a iframe. Will explore in next commit.