Add focused error state to character count
If a user goes over the character limit whilst typing, their only feedback for having done so is in the count message appear below the textarea (for a sighted user) or in the buffered announcement (for a screen reader user).

This means that users on small screens or using high magnification settings may not be aware that they have gone past the character limit until they have left the textarea and scrolled down. A sighted user may also simply not notice the character count text changing to alert them of the issue.

It was suggested in #1908 that the character count should maintain having the error state border even when the field has been given focus, so that there is another mechanism for users to recognise the issue. This diverges from the standard for other input types, however, as the character count is the only form control component that performs 'live' validation, the exception may make sense here.