Fix disabled File upload hover state
Avoid visual changes when hovering the `<label>`, which triggers the associated field's `:hover` state.

Moves the button's disabled state to the end of the component CSS so it can override any of the other states,
and uses variables to reset the colour to the empty state colours (as a file input cannot be pre-filled).