Create a separate route for '/components'
The `/components` route was not rendering any example because the `.param('componentName')` handler
is only run if the param is present in the URL (which it is not if the path is just `/components`).

Instead, this creates a separate route for `/components`, where the state is assigned to each components fixtures
in the route handler.