Only assign state to fixture that will actually be used
`.param(componentName)` only needed to assign the state to the fixtures of all components
because the `/components/:componentName?` route was also handling `/components`.

Now it doesn't we can clean this up a little and only assign the state to the component
whose fixtures are being rendered, rather than all components.