Avoid injecting the states in the fixtures
Discovered that macros could receive the context they're used in through this comment: https://github.com/mozilla/nunjucks/issues/1020#issuecomment-320206817. This allows us to set the states to render as examples in `res.locals` and not modify the fixtures we read from `govuk-frontend`, limiting the risks of collisions.

If Nunjucks behaves like Jinja, seems we're losing some caching of the compiled `showExamples` macro. However, given we're not having high requirements in response times, I think we're good to do that 😊