Remove computation of example in `componentName` param
Following th path of creating a separate route when a param of the route is optional,
as it reduces the complexity of the route handler and makes it easier to figure which
route a specific URL triggers.

This extract a custom route for rendering the default example when hitting `/components/:componentName/preview`,
which rewrites the URL and lets the handler of `/components/:componentName/:exampleName/preview` do the actual
rendering.

In turn, this avoids to the handler of `.param('componentName')` the worry of assigning a default example to the locals,
in case there's no `:exampleName` in the path. That responsibility is now solely on the handler of `.param('exampleName')`,
which makes both handlers do only what they say on the tin.