Remove computation of example in `componentName` param
This was only necessary because `:exampleName` was an optional parameter. By making it explicit and adding some URL rewriting when it's missing, we can focus the `componentName` param to only handling concepts related to the component not a default example.

This both provides clearer encapsulation, as well as avoids looking up an example when we don't need one.