Use `filter` to recolour the crown copyright logo
We wanted to change the crown copyright logo colour but didn't want to
introduce a breaking change. Therefore it has been split into a
pseudoelement with a `filter` applied to change the colour. This means
that filter to change the colour isn't applied to the text of the link,
just the logo.

The CHANGELOG has also been updated.