Remove classes we don't use from logo macro
This lightens the markup as we don't need these classes for styling.
To account for that, tests have been updated to target SVG elements rather than classes,
using:
- the presence of a `g` tag as a marker of the Tudor crown
- the presence of a `path` after the `g` tag as a marker of the logotype
- the presence of a `circle` as a sibling of the `g` tag as a marker of the rebranded logo

As we're not necessarily planning to update that markup often (if at all),
I think it's fine to have selectors tightly coupled to the structure of the markup here.
Especially as it provides a more explicit approach to what we're looking for than `toMatchSnapshot`
while allowing us to only ship the markup we need.