Use logo macro in footer
If `rebrand` is set, display the crown logo in the footer above the
meta section. Uses the logo macro with the logotype disabled.

Add tests for the `rebrand` option of the `govukFooter` macro

Adds a `withRebrand` helper to easily add the `rebrand: true` option to existing examples,
and tests to check when the crown is displayed.

Document `rebrand` Nunjucks macro option

Add CHANGELOG entry for the crown addition to the Footer

Describes how to enable it both when using the macro and updating the HTML.

Reference to the 'feature flag described in the previous section' is copied from
the Header update, and will likely need updating when reviewing the CHANGELOG
once all the content is present, referring back to the change to the Page template.