Ignore HTML files in `dist`
Adds a negative glob pattern in the `files` field of `package.json`.
This allows us to exclude the HTML files in `dist` without needing
to remove `dist` from `.gitignore` (which could possibly cause issues
down the line of us commiting things we don't want to commit should
we miss something when updating `.gitignore`).