Update generateReleaseNotes to check for an internal identifier
Firstly changes the flow of the function to take the version being passed to it, which we'll always have either via the validated input or the GH tag.

Secondly updates getChangelogLineIndexes to build the regex it passes to findIndexOfFirstMatchingLine in a more sensible way, separating the baseline heading regex into its own function which we pass processed heading regex to (buildHeadingRegexQuery).