Create macro to compute whether the rebrand is active
Enables components to share the same logic when assessing whether
they should render adjustments for the brand refresh.

1. The `govukSetRebrand` macro computes whether the rebrand is enabled
based on the `params` received by the macro and the `govukRebrandGlobal`
2. The components use a `{% call %}` block with an argument to call the macro
and receive whether the rebrand is active or not, allowing them to react accordingly

As this is syntactic sugar to avoid repetition across our components,
it feels fine to use advanced Nunjucks features which may be tricky to replicate
in other templating languages.