Omit ‘empty’ nav items from service navigation
When defining the list of navigation items, users might opt to selectively include certain items depending on a condition, for example the user being logged in or having certain permissions.

For example, you might omit an ‘admin’ link conditionally like this:

```
{% set showAdmin = "false" %}

{{
  govukServiceNavigation({
    navigation: [
      {
        href: "/reports",
        text: "View reports"
      },
      {
        href: "/admin",
        text: "Admin"
      } if showAdmin,
      {
        href: "/roles",
        text: "Roles"
      }
    ]
  })
}}
```

These entries are still present in the `params.navigation` array when the condition is not met but are falsey. When looping over `params.navigation` we still include an empty `<li>` for this falsey entry. As these list items have a margin, this also results in a gap appearing in the navigation where the conditional item would have appeared.

Filter out the falsey values using the Nunjucks `select` filter. We do this before we check if any navigation items are present so we can omit the navigation entirely if all navigation items are falsey.

Closes #5862