Ensure logo always uses currentcolor
Default to matching the text colour from the logo’s context rather than the default fill colour of black.

This means that the crown will appear in `#0b0c0c` rather than `#000000` in the footer, but also that it will match the user’s preferred text colour when in forced colors mode.