Remove fill from SVG
Now that we’re using the SVG as a mask we don’t need it to have colour information. It’s also misleading as the SVG renders in a colour that is not displayed anywhere on the site.

Finally this means that in browsers that don’t support mask-image, where we fall back to background-image, the crest will be rendered in full black which is closer to how we use it when masked than the light grey that was used previously.