Add full page example of a product page
The product pages (for example for GOV.UK Pay) make extensive use of blue on their homepage, specifically by using a large blue masthead at the top of the page.

Add an example of a product page based on Pay’s so we can see how it works with the recently released service navigation component, as well as the upcoming brand changes.

In order to make it realistic I’ve chosen to include the top 2 images used on Pay’s current product page. The Express app does not currently serve any images, so I’ve added a new route which uses `res.sendFile` to serve images relative to the view.

I’m not 100% sure that having the images under the views folder makes sense. I thought it was better to keep everything together in one place, but we do already have the stylesheets stored under stylesheets/full-page-examples.