Test for `mask-position` rather than `mask-image`
Edge 17 seems to pass the check `@supports (mask-image: url(“”))` but displays a black box instead of the SVG. This is likely because Edge 17 does not support SVGs for mask images [1].

Test instead for `mask-position` which correctly fails in Edge 17, meaning we fall back to using the `background-image` approach.

According to caniuse [2] [3], the only support in difference between these two properties is:

- support for `mask-position` begins with Edge 18, rather than Edge 16 for `mask-image`
- support for `mask-position` begins with Safari 4, rather than Safari 3.1 for `mask-image`

This also feels like the more correct test given that we rely on `mask-position` working for the mask to display correctly.

[1]: https://caniuse.com/mdn-css_properties_mask-image_svg_masks
[2]: https://caniuse.com/mdn-css_properties_mask-image
[3]: https://caniuse.com/mdn-css_properties_mask-position