Fix footer licence link reflowing on focus
Safari does not currently support `text-wrap: balance` alongside `box-decoration-break: clone` [1]. This has the odd effect at the minute that when the license link in the footer is focused it reflows, as though `text-wrap: balance` has been disabled on its parent.

We’ve been using `box-decoration-break: clone` on focused links since 05ed7c9 to work around an issue introduced in Chromium 108, where the box-shadow was missing from the focus state for links split over multiple lines.

That issue was fixed in Chromium 112. Unfortunately because Chromium 109 was the last version to work with Windows 7, Windows Server 2008 R2, Windows 8, Windows Server 2012, Windows 8.1 and Windows Server 2012 R2 there’s a peak of users stuck on this version of Chromium [2], so we ideally want to keep the fix in place for the time-being.

As Chromium didn’t introduce support for `text-wrap: balance` until 114 we can use it to guard the `box-decoration-break: clone` property.

Chromium 108-111 will continue to use `box-decoration-break: clone` for focused links. Newer browsers that support `text-wrap: balance` will not, avoiding the conflict.

[1]: https://bugs.webkit.org/show_bug.cgi?id=260351
[2]: https://github.com/alphagov/govuk-frontend/issues/5878#issuecomment-2838369733