Add renderMacro function for forward compatibility
The renderMacro function is a slightly tweaked version of the renderMacro method that we use in v5, which we replaced callMacro with.

To allow us to copy test files (mostly) wholesale, we can create a v4 version of this function. Otherwise, we'd have to use callMacro  and adjust the test files to match the parameters and it'd be fiddly.