Fix govuk-font-size mixin in libsass
We’re using interpolation to try and check whether a given variable matches any of the keys of the `$govuk-typography-scale map` when both the variable and the keys are treated as strings.

Unfortunately in node sass the interpolation results in the entire expression being treated as a string, rather than each side of the expression being interpolated into a separate string and then compared.

In other words, it’s equivalent to us writing `@if not $font-map and "#{$font-size} == #{$size}”` if we were using quoted strings [1].

Because a string is truthy, the `@if` statement evaluates as true and $font-map is set to whatever the first entry in the map is, regardless of what `$size` is.

This means that the `govuk-font-size` mixin is currently outputting the font properties for size 80 in our typography scale when passed a `$size` that does not map directly to a key from the map.

Quote both values so that they are correctly treated as separate strings and then compared.

There’s no point adding a test for this as all of our tests run under Dart Sass where this problem doesn’t exist.

[1]: We can see this by testing with some debug statements:

```
@debug(true == false);
@debug(#{true} == #{false});
@debug("#{true} == #{false}");
@debug("#{true}" == "#{false}");
```

With Dart Sass, you get:
test.scss:1 Debug: false
test.scss:2 Debug: false
test.scss:3 Debug: true == false
test.scss:4 Debug: false

With Node Sass, you get:
test.scss:1 DEBUG: false
test.scss:2 DEBUG: true == false
test.scss:3 DEBUG: true == false
test.scss:4 DEBUG: false