Ignore scss as a highlight language for typedoc
Typedoc v0.28.3 made a change so that only certain languages are loaded into the highlighter by default [1].

We have an `scss` example in our README which is included as the initial screen in the generated typedoc. `scss` isn’t included in these default languages which means typedoc is error’ing with:

```
[error] Language `scss` not found, you may need to load it first
[error] html output could not be generated due to the errors above
[error] Found 2 errors and 24 warnings
```

Prevent this by adding `scss` to the list of ignored highlight languages [2]. Alternatively we could have set `highlightLanguages`, but it could go out of sync and we don't care about nicely highlighted scss in our JS docs.

[1]: https://github.com/TypeStrong/typedoc/issues/2943
[2]: https://typedoc.org/documents/Options.Output.html#ignoredhighlightlanguages