Simplify prefix for breakpoint custom properties
Since #4562 we’ve exposed our breakpoints (mobile, tablet, desktop) as CSS custom properties so that we can use them in `matchMedia` functions in our JavaScript without having to duplicate the values.

When we introduced them we prefixed them with `--govuk-frontend`, I think to align with the existing `--govuk-frontend-version` property.

But the `--govuk-frontend-version` property is _about_ GOV.UK Frontend, whereas I think we can simplify the prefix for the rest of our custom properties to just `--govuk`, matching what we do in Sass for our functions, mixins and variables.

Keep the existing custom property names as aliases just in case anyone else is using them, but mark them as deprecated to be removed in the next major release.