Don’t use menu for service nav with a single link
If the service navigation component only includes a single navigation item, it doesn’t really make sense to hide that item inside a menu on mobile as it doesn’t save any space compared to just showing the single link.

Give users control over this by introducing a new `collapseNavigationOnMobile` param, which defaults to true if there is more than one navigation item, otherwise false.

Only include the menu button if `collapseNavigationOnMobile` is true.

The JavaScript already exits early if the menu button is not present.